Split tool card surface models
