+++ docs/DECISIONS.md
@@ -141,6 +141,7 @@
 - Desktop app bootstrap lives in `QuillCodeDesktopApp.swift` and should remain declarative. Native commands, menu-bar layout, OAuth loopback capture, browser fetching, automation notifications, and controller orchestration live in separate desktop files. Parity gates intentionally scan the whole `Sources/quill-code-desktop` folder so extraction can continue without pushing implementation details back into scene composition.
 - Desktop controllers should route UI and apply workspace state, not own protocol flows or cancellable task bookkeeping. TrustedRouter sign-in lives in `QuillCodeDesktopSignInCoordinator`, and send/terminal/browser/automation task slots live in `QuillCodeDesktopTaskCoordinator`. Parity gates check those boundaries so OAuth exchange and raw task IDs do not drift back into `QuillCodeDesktopController.swift`.
 - Desktop settings persistence lives in `QuillCodeDesktopSettingsCoordinator`. The controller applies settings results and runtime refreshes, while the coordinator owns TrustedRouter key replacement/clear rules, OAuth account reset rules, and config persistence. macOS Computer Use System Settings URLs live in `MacSystemSettingsOpener`, keeping platform URL routing out of the controller and out of settings persistence.
+- Tool-card and artifact presentation models live in `QuillCodeToolCardSurface.swift`, not in `WorkspaceModel.swift`. `WorkspaceModel` may assemble tool cards from events, but status/density, artifact kind detection, document/image/text-preview metadata, and default expansion behavior are surface concerns with their own parity gate.
 - Workspace command IDs reduce through `WorkspaceCommandPlan` before touching `WorkspaceModel` side effects. The planner owns prefix parsing, draft-prefill mapping, static action mapping, automation quick recurrence parsing, and canonical command-ID-to-tool-name mapping, while `WorkspaceModel` owns mutations and tool dispatch. This keeps command palette rows, slash-command templates, automation row actions, MCP lifecycle actions, memory actions, and git shortcuts from drifting as the command set grows.
 - Automation record construction and due-run draft creation live in `WorkspaceAutomationEngine.swift`, not inside `WorkspaceModel`. `WorkspaceAutomationFactory` owns concrete `QuillAutomation` records and schedule helpers, `WorkspaceAutomationRunner` owns due filtering, recurrence advancement, and generated thread drafts, and `WorkspaceModel` applies those drafts to UI selection, project refresh, stores, and notification reports. This keeps scheduler behavior testable without pulling project persistence or top-bar state into pure automation tests.
 - Multi-agent collaboration should happen through pull requests and a serialized merge train rather than direct pushes to `main`. CI runs on `pull_request`, `push`, and `merge_group` events. The train watches `merge-train` and `automerge` labels, ignores draft or `do-not-merge` PRs, processes only the oldest eligible PR, updates behind branches before merging, and merges only after all non-train checks are successful. This keeps Codex/agent work reviewable while preserving a green mainline.
